+++ backend/teluslibrary/src/main/java/com/appllication/teluslibrary/repositories/LoanRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.appllication.teluslibrary.entities.Loan;
 
-public interface LoanRepository extends JpaRepository<Loan, Integer>{
+public interface LoanRepository extends JpaRepository<Loan, Long>{
 
 }
